Put both operator sets in one file
The merge left the awaiting half in a partial across a second file, which kept
the split it was meant to remove. ClrEnumerableDefaults is one class in one file
now, pulled operators then awaiting ones, the same shape ClrBuiltInMethod already
has: a sync block, then an async block, one suffix apart.

The two class remarks are folded into one. Two sentences in the awaiting half's
had to change rather than move, because they described a separate class and a
separate file that no longer exist.
